refactor: move item effect display to detail card
Item effects are now shown in the BuySellPanel detail card instead
of cluttering the compact item buttons in ShopInventory.
