+++ src/components/shop/ShopInventory.tsx
@@ -2,19 +2,8 @@
  * Shop inventory component displaying items available for purchase.
  */
 
-import { toDisplay, toDisplayCare } from "@/game/types/common";
 import type { Rarity } from "@/game/types/constants";
 import type { Item } from "@/game/types/item";
-import {
-  isBattleItem,
-  isCleaningItem,
-  isDrinkItem,
-  isEquipmentItem,
-  isFoodItem,
-  isMaterialItem,
-  isMedicineItem,
-  isToyItem,
-} from "@/game/types/item";
 import type { ShopItem } from "@/game/types/shop";
 import { cn } from "@/lib/utils";
 
@@ -47,63 +36,6 @@ function getRarityClass(rarity: Rarity): string {
 }
 
 /**
- * Format a care stat value as a range if floor and ceil differ.
- * Shows `+(min)~(max)` if there's a difference, otherwise just `+(value)`.
- */
-function formatCareStatRange(microValue: number): string {
-  const min = toDisplay(microValue);
-  const max = toDisplayCare(microValue);
-  if (min === max) {
-    return `+${min}`;
-  }
-  return `+${min}~${max}`;
-}
-
-/**
- * Get the effect text for an item based on its category.
- */
-function getItemEffect(item: Item): string | null {
-  if (isFoodItem(item)) {
-    return `${formatCareStatRange(item.satietyRestore)} Satiety`;
-  }
-  if (isDrinkItem(item)) {
-    const parts = [`${formatCareStatRange(item.hydrationRestore)} Hydration`];
-    if (item.energyRestore) {
-      parts.push(`+${toDisplay(item.energyRestore)} Energy`);
-    }
-    return parts.join(", ");
-  }
-  if (isToyItem(item)) {
-    return `${formatCareStatRange(item.happinessRestore)} Happiness`;
-  }
-  if (isCleaningItem(item)) {
-    return `Removes ${item.poopRemoved} poop`;
-  }
-  if (isMedicineItem(item)) {
-    const parts: string[] = [];
-    if (item.isFullRestore) {
-      parts.push("Full HP restore");
-    } else if (item.healAmount) {
-      parts.push(`+${item.healAmount} HP`);
-    }
-    if (item.cureStatus && item.cureStatus.length > 0) {
-      parts.push(`Cures: ${item.cureStatus.join(", ")}`);
-    }
-    return parts.length > 0 ? parts.join(", ") : null;
-  }
-  if (isBattleItem(item)) {
-    return `+${item.modifierValue}% ${item.statModifier} (${item.duration} turns)`;
-  }
-  if (isEquipmentItem(item)) {
-    return item.effect;
-  }
-  if (isMaterialItem(item)) {
-    return "Crafting material";
-  }
-  return null;
-}
-
-/**
  * Displays a grid of items available for purchase in a shop.
  */
 export function ShopInventory({
@@ -125,7 +57,6 @@ export function ShopInventory({
       {items.map(({ shopItem, itemDef }) => {
         const isSelected = selectedItemId === shopItem.itemId;
         const canAfford = playerCoins >= shopItem.buyPrice;
-        const itemEffect = getItemEffect(itemDef);
 
         return (
           <button
@@ -144,11 +75,6 @@ export function ShopInventory({
             <span className="text-xs font-medium truncate max-w-full">
               {itemDef.name}
             </span>
-            {itemEffect && (
-              <span className="text-xs text-muted-foreground truncate max-w-full">
-                {itemEffect}
-              </span>
-            )}
             <span
               className={cn(
                 "text-xs",
